requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
astroid==2.2.5
cassandra-driver==3.17.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
